% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopiapi-package.R
\docType{package}
\name{ecopiapi-package}
\alias{ecopiapi}
\alias{ecopiapi-package}
\title{ecopiapi: R package to interact with the ecopi API}
\description{
Get, post, patch data from the ecop API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OekoFor/ecopiapi}
  \item Report bugs at \url{https://github.com/OekoFor/ecopiapi/issues}
}

}
\author{
\strong{Maintainer}: Felix Guenther \email{guenther@oekofor.de}

}
\keyword{internal}
