% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{get_api_key}
\alias{get_api_key}
\title{Retrieve API key from environment variables}
\usage{
get_api_key(api_key_name = "API_KEY")
}
\arguments{
\item{api_key_name}{A character string specifying the name of the API key. Default is "API_KEY".}
}
\value{
A character string representing the value of the API key.
}
\description{
This function retrieves the value of an API key from the environment variables by its name.
}
\examples{
\dontrun{
  # Retrieve the value of an API key named 'API_KEY'
  api_key_value <- get_api_key("API_KEY")
}
}
