% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_historicalprojects}
\alias{get_historicalprojects}
\title{Get historical projects list.}
\usage{
get_historicalprojects(..., get_response = FALSE)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/docs/#tag/v0.2/operation/v0.2_historicalprojects_list}}

\item{get_response}{Logical. If TRUE, returns a list containing both the original response and the data frame.}
}
\value{
If get_response is FALSE (default), returns a data frame containing the historical projects.
If get_response is TRUE, returns a list with 'data' (the data frame) and 'response' (the original API response).
}
\description{
Wrapper around the 'historicalprojects_list' endpoint to retrieve a list of historical projects based on the specified query parameters.
}
\examples{
# retrieve a dataframe of all historical projects
\dontrun{
get_historicalprojects()
}

# Retrieve a list of projects that contain 'red_panda' or 'green_banana' in their name
\dontrun{
get_historicalprojects(project_name__in = c("pam_in_chemnitz"))
}

}
