% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_historicalrecordergroups}
\alias{get_historicalrecordergroups}
\title{Get historical recorder groups list.}
\usage{
get_historicalrecordergroups(...)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/docs/#tag/v0.1/operation/v0.1_historicalrecordergroups_list}.}
}
\value{
A list containing the historical recorder groups that match the specified query parameters: \url{https://api.ecopi.de/api/docs/#tag/v0.1/operation/v0.1_historicalrecordergroups_list}.
}
\description{
Wrapper around the 'historicalrecordergroups_list' endpoint to retrieve a list of recorder groups based on the specified query parameters.
This contains the configurations and species list
}
\examples{
# Retrieve a data frame of historical recorder groups for project 'oekofor'
\dontrun{
get_historicalrecordergroups(project_name = "oekofor")
}

}
