% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_historicalrecorders}
\alias{get_historicalrecorders}
\title{Get historical recorders list.}
\usage{
get_historicalrecorders(..., get_response = FALSE)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/docs/#tag/v0.2/operation/v0.2_historicalrecorders_list}.}

\item{get_response}{Logical. If TRUE, returns a list containing both the original response and the data frame.}
}
\value{
If get_response is FALSE (default), returns a data frame containing the historical recorders.
If get_response is TRUE, returns a list with 'data' (the data frame) and 'response' (the original API response).
}
\description{
Wrapper around the 'historicalrecorders_list' endpoint to get a list of recorders based on the specified query parameters.
}
\examples{
# Retrieve a list of historical recorders for project '017_neeri'
\dontrun{
get_historicalrecorders(project_name = "pam_in_chemnitz")
}

}
