% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_detections}
\alias{get_latest_detections}
\title{Get latest detections.}
\usage{
get_latest_detections(latest = 100)
}
\arguments{
\item{latest}{A numeric value indicating the number of latest detections to return (default 100).}
}
\value{
A tibble containing the latest detections.
}
\description{
Wrapper around the 'detections_read' endpoint to retrieve the latest detections.
}
\examples{
# Get the latest 100 detections
get_latest_detections()

# Get the latest 50 detections
get_latest_detections(50)
}
