% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_detections_by_project}
\alias{get_latest_detections_by_project}
\title{Get latest detections by project name.}
\usage{
get_latest_detections_by_project(latest = 100, project_name)
}
\arguments{
\item{latest}{A numeric value indicating the number of latest detections to return (default 100).}

\item{project_name}{A character value indicating the name of the project to retrieve detections for.}
}
\value{
A tibble containing the latest detections for the specified project.
}
\description{
Wrapper around the 'detections_project_read' endpoint to retrieve the latest detections for a specific project.
}
\examples{
\dontrun {
#' # Get the latest 100 detections for a project named 'my_project'
get_latest_detections_by_project(project_name = 'my_project')
}

# Get the latest 50 detections for a project named 'my_project'
get_latest_detections_by_project(latest = 50, project_name = "my_project")
}
