% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_detections_by_recorder}
\alias{get_latest_detections_by_recorder}
\title{Get latest detections by recorder name.}
\usage{
get_latest_detections_by_recorder(latest = 100, recorder_name)
}
\arguments{
\item{latest}{A numeric value indicating the number of latest detections to return (default 100).}

\item{recorder_name}{A character value indicating the name of the recorder to retrieve detections for.}
}
\value{
A tibble containing the latest detections for the specified recorder.
}
\description{
Wrapper around the 'detections_recorder_read' endpoint to retrieve the latest detections for a specific recorder.
}
\examples{
# Get the latest 100 detections for a recorder named 'my_recorder'
get_latest_detections_by_recorder(recorder_name = "my_recorder")

# Get the latest 50 detections for a recorder named 'my_recorder'
get_latest_detections_by_recorder(latest = 50, recorder_name = "my_recorder")
}
