% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_detections_by_recordergroup}
\alias{get_latest_detections_by_recordergroup}
\title{Get latest detections by recorder group.}
\usage{
get_latest_detections_by_recordergroup(
  latest = 100,
  project_name,
  recordergroup_name = "default"
)
}
\arguments{
\item{latest}{A numeric value indicating the number of latest detections to return (default 100).}

\item{project_name}{A character value indicating the name of the project to retrieve detections for.}

\item{recordergroup_name}{A character value indicating the name of the recorder group to retrieve detections for. Default is 'default'.}
}
\value{
A tibble containing the latest detections for the specified recorder group.
}
\description{
Wrapper around the 'detections_recordergroup_read' endpoint to retrieve the latest detections for a specific recorder group.
}
\examples{
# Get the latest 100 detections for the default recorder group in project 'my_project'
get_latest_detections_by_recordergroup(project_name = "my_project")

# Get the latest 50 detections for a recorder group named 'my_group' in project 'my_project'
get_latest_detections_by_recordergroup(latest = 50, project_name = "my_project", recordergroup_name = "my_group")
}
