% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_recorder_states}
\alias{get_latest_recorder_states}
\title{Get latest recorder states.}
\usage{
get_latest_recorder_states(recorder_name, latest = 100)
}
\arguments{
\item{recorder_name}{The name of the recorder to retrieve the latest states for.}

\item{latest}{The number of latest recorder states to retrieve (default is 100).}
}
\value{
A tibble containing the latest recorder states for the specified recorder: \url{https://api.ecopi.de/api/docs/#operation/recorderstates_recorder_read}.
}
\description{
Wrapper around the 'recorderstates_recorder_read' endpoint to retrieve the latest recorder states for a specific recorder.
}
\details{
This function retrieves the latest recorder states for a specific recorder, based on the recorder name provided in the 'recorder_name' parameter.
The available information about recorder states is documented in the EcoPi API documentation: \url{https://api.ecopi.de/api/docs/#operation/recorderstates_recorder_read}.
}
\examples{
# Retrieve the latest 100 recorder states for recorder '00000000d76d0bf9'
get_latest_recorder_states(recorder_name = "00000000d76d0bf9")

# Retrieve the latest 50 recorder states for recorder '00000000d76d0bf9'
get_latest_recorder_states(recorder_name = "00000000d76d0bf9", latest = 50)

}
