% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_mediafile}
\alias{get_mediafile}
\title{Get media file.}
\usage{
get_mediafile(uid)
}
\arguments{
\item{uid}{uid of that specifiy detection. See \url(https://api.ecopi.de/api/v0.1/docs/#operation/detections_media_read)}
}
\value{
object of type "httr2_response". contains raw data in body (currently audio or image)
}
\description{
Wrapper around the 'MediaFile Detection' endpoint to retrieve a media file from a detection.
}
\examples{
get_mediafile("c8c155f9-c05b-4e86-b842-88b80829e36c")

}
