% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_project_info}
\alias{get_project_info}
\title{Get project info.}
\usage{
get_project_info(project_name, get_response = FALSE)
}
\arguments{
\item{project_name}{Name of the project.}

\item{get_response}{Logical. If TRUE, returns a list containing both the original response and the data frame.}
}
\value{
If get_response is FALSE (default), returns a list containing information about the specified project.
If get_response is TRUE, returns a list with 'data' (the project info) and 'response' (the original API response).
}
\description{
Wrapper around the 'projects_retrieve' endpoint to retrieve information about a specific project.
}
\details{
This function retrieves information about a specific project, based on the project name provided in the 'project_name' parameter.
}
\examples{
# Retrieve information about the '017_neeri' project
\dontrun{
get_project_info(project_name = "017_neeri")
}

}
