% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_projects_list}
\alias{get_projects_list}
\title{Get projects list.}
\usage{
get_projects_list(params = list())
}
\arguments{
\item{params}{A list of key-value pairs representing the query parameters to be sent with the API request.}
}
\value{
A tibble containing the projects that match the specified query parameters: \url{https://api.ecopi.de/api/docs/#operation/projects_list}.
}
\description{
Wrapper around the 'projects_list' endpoint to retrieve a list of projects based on the specified query parameters.
}
\details{
The 'params' parameter is a list of key-value pairs that represent the query parameters to be sent with the API request.
Each key represents a query parameter, and each value represents the value of the query parameter.
For example, to retrieve projects that contain a specific string in their name, you can set the 'name__contains' parameter to the string to search for.
The available query parameters are documented in the EcoPi API documentation: \url{https://api.ecopi.de/api/docs/#operation/projects_list}.
}
\examples{
# retrieve a list of all projects
get_projects_list()

# Retrieve a list of projects that contain 'red_panda' or 'green_banana' in their name
get_projects_list(params = list("project_name__in" = "red_panda, green_banana"))

}
