% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recorder_log_info}
\alias{get_recorder_log_info}
\title{Get recorder log info.}
\usage{
get_recorder_log_info(recorder_name)
}
\arguments{
\item{recorder_name}{The name of the recorder to retrieve information about.}
}
\value{
A list containing information about the specified recorder log: \url{https://api.ecopi.de/api/docs/#operation/recorderlogs_read}.
}
\description{
Wrapper around the 'recorderlogs_read' endpoint to retrieve information about a specific recorder log.
}
\details{
This function retrieves information about a specific recorder log, based on the recorder name provided in the 'recorder_name' parameter.
The available information about a recorder log is documented in the EcoPi API documentation: \url{https://api.ecopi.de/api/docs/#operation/recorderlogs_read}.
}
\examples{
# Retrieve information about the recorder log for recorder '00000000d76d0bf9'
get_recorder_log_info(recorder_name = "00000000d76d0bf9")

}
