% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recorder_states_list}
\alias{get_recorder_states_list}
\title{Get recorder states list.}
\usage{
get_recorder_states_list(params = list())
}
\arguments{
\item{params}{A list of key-value pairs representing the query parameters to be sent with the API request.}
}
\value{
A tibble containing the recorder states that match the specified query parameters: \url{https://api.ecopi.de/api/docs/#operation/recorderstates_list}.
}
\description{
Wrapper around the 'recorderstates_list' endpoint to retrieve a list of recorder states based on the specified query parameters.
}
\details{
The 'params' parameter is a list of key-value pairs that represent the query parameters to be sent with the API request.
Each key represents a query parameter, and each value represents the value of the query parameter.
For example, to retrieve recorder states for a specific recorder, you can set the 'recorder_name' parameter to the name of the recorder.
The available query parameters are documented in the EcoPi API documentation: \url{https://api.ecopi.de/api/docs/#operation/recorderstates_list}.
}
\examples{

# Retrieve all recorder states
get_recorder_states_list()

# Retrieve a list of recorder states for recorder '00000000d76d0bf9'
get_recorder_states_list(params = list("recorder_name" = "00000000d76d0bf9"))

}
