% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recordergroups_list}
\alias{get_recordergroups_list}
\title{Get recorder groups list.}
\usage{
get_recordergroups_list(params = list())
}
\arguments{
\item{params}{A list of key-value pairs representing the query parameters to be sent with the API request.}
}
\value{
A list containing the recorder groups that match the specified query parameters: \url{https://api.ecopi.de/api/docs/#operation/recordergroups_list}.
}
\description{
Wrapper around the 'recordergroups_list' endpoint to retrieve a list of recorder groups based on the specified query parameters.
This contains the configurations and species list
}
\details{
The 'params' parameter is a list of key-value pairs that represent the query parameters to be sent with the API request.
Each key represents a query parameter, and each value represents the value of the query parameter.
For example, to retrieve recorder groups for a specific project, you can set the 'project_name' parameter to the name of the project.
The available query parameters are documented in the EcoPi API documentation: \url{https://api.ecopi.de/api/docs/#operation/recordergroups_list}.
}
\examples{
# Retrieve a list of recorder groups for project 'oekofor'
get_recordergroups_list(params = list("project_name" = "oekofor"))

}
