% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recorderlogs}
\alias{get_recorderlogs}
\title{Get recorder logs list.}
\usage{
get_recorderlogs(..., get_response = FALSE)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/docs/#tag/v0.2/operation/v0.2_recorderlogs_list}.}

\item{get_response}{Logical. If TRUE, returns a list containing both the original response and the data frame.}
}
\value{
If get_response is FALSE (default), returns a data frame containing the recorder logs.
If get_response is TRUE, returns a list with 'data' (the data frame) and 'response' (the original API response).
}
\description{
Wrapper around the 'recorderlogs_list' endpoint to retrieve a list of recorder logs based on the specified query parameters.
}
\examples{
# Retrieve a dataframe of recorder logs for project '017_neeri'
\dontrun{
get_recorderlogs(project_name = "pam_in_chemnitz")
}

}
