% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recorders_count_detections_project_recorder_recordings}
\alias{get_recorders_count_detections_project_recorder_recordings}
\title{Get a count of detections per recorder grouped by rerecording.}
\usage{
get_recorders_count_detections_project_recorder_recordings(
  project_name,
  recorder_field_id,
  ...
)
}
\arguments{
\item{project_name}{Name of the project to get a count on detetcions class and recorder}

\item{recorder_field_id}{Recorder field ID}

\item{...}{query paramaters. See \url{https://api.ecopi.de/api/docs/#tag/v0.2/operation/v0.2_aggregations_projects_recorders_recordings_count_detections_retrieve}}
}
\value{
A summary containing species counts per recorder per recording within a project that match the specified query parameters: \url{https://api.ecopi.de/api/docs/#tag/v0.2/operation/v0.2_aggregations_projects_recorders_recordings_count_detections_retrieve}.
}
\description{
Wrapper around the 'aggregations_project_recorders_count_detections_project_recorder_recordings' endpoint to retrieve a count of detections for each recorder and grouped by recording.
}
\examples{
# Retrieve a count of detections for each ecoPi at each recording. (Project x , with 3 ecoPi'S is scheduled to record at 10 times per day. I want to extract a count of detections for each ecoPi and each scheduled recording session )
\dontrun{
get_recorders_count_detections_project_recorder_recordings(project_name = "pam_in_chemnitz", recorder_field_id = 1)
}

}
