% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recorders_list}
\alias{get_recorders_list}
\title{Get recorders list.}
\usage{
get_recorders_list(params = list())
}
\arguments{
\item{params}{A list of key-value pairs representing the query parameters to be sent with the API request.}
}
\value{
A list containing the recorders that match the specified query parameters: \url{https://api.ecopi.de/api/docs/#operation/recorders_list}.
}
\description{
Wrapper around the 'recorders_list' endpoint to retrieve a list of recorders based on the specified query parameters.
}
\details{
The 'params' parameter is a list of key-value pairs that represent the query parameters to be sent with the API request.
Each key represents a query parameter, and each value represents the value of the query parameter.
For example, to retrieve recorders for a specific project, you can set the 'project_name' parameter to the name of the project.
The available query parameters are documented in the EcoPi API documentation: \url{https://api.ecopi.de/api/docs/#operation/recorders_list}.
}
\examples{
# Retrieve a list of recorders for project '017_neeri'
get_recorders_list(params = list("project_name" = "017_neeri"))

}
