% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recorderstates}
\alias{get_recorderstates}
\title{Get recorder states list.}
\usage{
get_recorderstates(...)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/v0.1/docs/#operation/recorderstates_list}}
}
\value{
A dataframe containing the recorder states that match the specified query parameters: \url{https://api.ecopi.de/api/docs/#operation/recorderstates_list}.
}
\description{
Wrapper around the 'ListView RecorderStates' endpoint to retrieve a list of recorder states based on the specified query parameters.
}
\examples{
# Retrieve all recorder states
get_recorderstates()

# Retrieve a list of recorder states for recorder '00000000d76d0bf9'
get_recorder_states(recorder_name = "00000000d76d0bf9")

}
