% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{patch_detections}
\alias{patch_detections}
\title{PATCH detection}
\usage{
patch_detections(..., id_or_uid, new_data)
}
\arguments{
\item{...}{query paramaters. See \url(https://api.ecopi.de/api/v0.1/docs/#operation/detections_list)}
}
\value{
httr2_response
}
\description{
Wrapper around the 'ListView Detections' endpoint to update detections parameters based on the specified query parameters.
}
\examples{
# Update the parameter confirmed of an example detection
patch_detections(detection_id=123, datetime__month = 3)

}
