% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{patch_recordergroups}
\alias{patch_recordergroups}
\title{PATCH recordergroups}
\usage{
patch_recordergroups(..., id, new_data)
}
\arguments{
\item{...}{query paramaters. See \url(https://api.ecopi.de/api/v0.1/docs/#operation/detections_list)}
}
\value{
httr2_response
}
\description{
Wrapper around the 'DetailView RecorderGroup' endpoint to update recordergroups parameters based on the specified query parameters.
}
\examples{
# Update the parameter confirmed of an example detection
patch_recordergroups(id = "44", description="patch test")

}
