% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{patch_recorders}
\alias{patch_recorders}
\title{PATCH recorder.}
\usage{
patch_recorders(..., recorder_name, file_path)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/docs/#tag/v0.2/operation/v0.2_recorders_partial_update}.}

\item{recorder_name}{The name of the recorder to update information from.}

\item{file_path}{Path to file to upload (Watchout! If file_path given, new_data gets ignored (cannot patch multiple types)).}
}
\value{
httr2_response
}
\description{
Wrapper around the 'recorders_partial_update' endpoint to update recorders parameters based on the specified body schema.
}
\examples{
# Update the parameter description of the recorder 00041aefd7jgg1014
\dontrun{
patch_recorders(recorder_name = "008041aefd7ee1015", description = "This a recorder ...", lat = 48)
}
# OR with image
\dontrun{
patch_recorders(recorder_name = "00041aefd7jgg1014", description = "Teeeest ...", lat = 48, file_path = "/sample_path/sample.jpeg")
}

}
