% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{resp_body_json_to_df}
\alias{resp_body_json_to_df}
\title{Convert response body to data frame}
\usage{
resp_body_json_to_df(api_response)
}
\arguments{
\item{api_response}{Response from API}
}
\value{
Data frame
}
\description{
Convert response body to data frame
}
\examples{
\dontrun{
response <- ecopi_api("GET /detections/", params = list("project" = "017_neeri"))
resp_body_json_to_df(response)
}
}
