% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API key as an environment variable}
\usage{
set_api_key(api_key_name = NULL, api_key = NULL)
}
\arguments{
\item{api_key_name}{A character string specifying the name of the API key.}

\item{api_key}{A character string specifying the value of the API key. If not provided, the user will be prompted to enter it.}
}
\value{
NULL. The function sets the API key as an environment variable and does not return a value.
}
\description{
This function sets an API key as an environment variable using a provided key name and key value.
If the key value is not provided, the user will be prompted to enter it.
}
\examples{
\dontrun{
  # Set an API key with a name 'API_KEY' and a value 'your_api_key'
  set_api_key(api_key_name = "API_KEY", api_key = "your_api_key")
}
}
