% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ecopi_api}
\alias{ecopi_api}
\title{Make API requests to the Ecopi API}
\usage{
ecopi_api(resource, ..., params = list(), new_data = list(), file_path)
}
\arguments{
\item{resource}{A character string specifying the API resource to request.}

\item{...}{Additional arguments to be passed to the req_template() function.}

\item{params}{A named list of query parameters to include in the request.}

\item{new_data}{A named list of parameters to be updated/ patched.}

\item{file_path}{A path currently only important to patch an image to a recorder endpoint.}
}
\value{
A \code{response} object from the httr2 package containing the API response.
}
\description{
This function sends an API request to the Ecopi API with the specified resource, parameters, and optional arguments.
}
\examples{
\dontrun{
# Send a request to the 'detections' endpoint. Only get detections in the project '017_neeri'
response <- ecopi_api("GET /detections/", params = list("project" = "017_neeri"))

#
}
}
