% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_by_url}
\alias{get_by_url}
\title{Get data from a linked URL endpoint (mostly provided from other gtes)}
\usage{
get_by_url(full_url, base_url = "https://api.ecopi.de/api/v0.2")
}
\arguments{
\item{full_url}{URL of linked endpoint}
}
\value{
Depends on the input. Mostly a data frame of the requested data.
}
\description{
Get data from a linked URL endpoint (mostly provided from other gtes)
}
\examples{
# Retrieve a ecopiapp release from a linked URL
\dontrun{
get_by_url("https://api.ecopi.de/api/v0.2/releases/69/")
}

}
