% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_detections_retrieve}
\alias{get_detections_retrieve}
\title{Get detections retrieve.}
\usage{
get_detections_retrieve(..., id_or_uid)
}
\arguments{
\item{...}{query paramaters. See \url(https://api.ecopi.de/api/docs/#tag/v0.1/operation/v0.1_detections_retrieve)}
}
\value{
A list containing the detection that match the specified query parameters: \url(https://api.ecopi.de/api/v0.1/docsco/#operation/detections_list)
}
\description{
Wrapper around the 'Retrieve Detections' endpoint to retrieve a single detection based on uid
}
\examples{
# Retrieve a single detection for a specific uid
get_detections_retrieve(uid="64733fbc-7cc8-49f6-adf1-c9ec2d676959")

}
