% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_historicalrecorders}
\alias{get_historicalrecorders}
\title{Get historical recorders list.}
\usage{
get_historicalrecorders(...)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/docs/#tag/v0.1/operation/v0.1_historicalrecorders_list}.}
}
\value{
A dataframe containing the historical recorders that match the specified query parameters: \url{https://api.ecopi.de/api/docs/#tag/v0.1/operation/v0.1_historicalrecorders_list}.
}
\description{
Wrapper around the 'historicalrecorders_list' endpoint to get a list of recorders based on the specified query parameters.
}
\examples{
# Retrieve a list of historical recorders for project '017_neeri'
get_historicalrecorders(project_name = "017_neerach_ried")

}
