% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_project_info}
\alias{get_project_info}
\title{Get project info.}
\usage{
get_project_info(project_name)
}
\arguments{
\item{project_name}{The name of the project to retrieve information about.}
}
\value{
A list containing information about the specified project: \url{https://api.ecopi.de/api/docs/#operation/projects_read}.
}
\description{
Wrapper around the 'projects_read' endpoint to retrieve information about a specific project.
}
\details{
This function retrieves information about a specific project, based on the project name provided in the 'project_name' parameter.
The available information about a project is documented in the EcoPi API documentation: \url{https://api.ecopi.de/api/docs/#operation/projects_read}.
}
\examples{
# Retrieve information about the '017_neeri' project
get_project_info(project_name = "017_neeri")

}
