% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_project_info}
\alias{get_project_info}
\title{Get project info.}
\usage{
get_project_info(project_name)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/docs/#tag/v0.1/operation/v0.1_projects_list}.}
}
\value{
A list containing information about the specified project: \url{https://api.ecopi.de/api/docs/#tag/v0.1/operation/v0.1_projects_list}.
}
\description{
Wrapper around the 'projects_retrieve' endpoint to retrieve information about a specific project.
}
\details{
This function retrieves information about a specific project, based on the project name provided in the 'project_name' parameter.
}
\examples{
# Retrieve information about the '017_neeri' project
get_project_info(project_name = "017_neeri")

}
