% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_projects}
\alias{get_projects}
\title{Get projects list.}
\usage{
get_projects(...)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/docs/#tag/v0.1/operation/v0.1_projects_list}}
}
\value{
A dataframe containing the projects that match the specified query parameters: \url{https://api.ecopi.de/api/docs/#tag/v0.1/operation/v0.1_projects_list}.
}
\description{
Wrapper around the 'projects_list' endpoint to retrieve a list of projects based on the specified query parameters.
}
\examples{
# retrieve a dataframe of all projects
\dontrun{
get_projects()
}

# Retrieve a list of projects that contain 'red_panda' or 'green_banana' in their name
\dontrun{
get_projects(project_name__in = "red_panda, green_banana")
}

}
