% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recorderlogs}
\alias{get_recorderlogs}
\title{Get recorder group info.}
\usage{
get_recorderlogs(...)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/v0.1/docs/#tag/recorderlogs}.}

\item{project_name}{The name of the project that the recorder group belongs to.}

\item{recordergroup_name}{The name of the recorder group to retrieve information about.}
}
\value{
A list containing information about the specified recorder group: \url{https://api.ecopi.de/api/docs/#operation/recordergroups_read}.

A dataframe containing the recorder logs that match the specified query parameters: \url{https://api.ecopi.de/api/v0.1/docs/#tag/recorderlogs}.
}
\description{
Wrapper around the 'recordergroups_read' endpoint to retrieve information about a specific recorder group.
}
\details{
This function retrieves information about a specific recorder group, based on the project name and recorder group name provided in the 'project_name' and 'recordergroup_name' parameters, respectively.
The available information about a recorder group is documented in the EcoPi API documentation: \url{https://api.ecopi.de/api/docs/#operation/recordergroups_read}.
}
\examples{
# Retrieve information about the 'default' recorder group for project 'oekofor'
get_recordergroup_info(project_name = "oekofor", recordergroup_name = "default")

# Retrieve a dataframe of recorder logs for project '017_neeri'
get_recorderlogs(project_name = "017_neeri")

}
