% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recorderlogs}
\alias{get_recorderlogs}
\title{Get recorder logs list.}
\usage{
get_recorderlogs(...)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/docs/#tag/v0.2/operation/v0.2_recorderlogs_list}.}
}
\value{
A dataframe containing the recorder logs that match the specified query parameters: \url{https://api.ecopi.de/api/docs/#tag/v0.2/operation/v0.2_recorderlogs_list}.
}
\description{
Wrapper around the 'recorderlogs_list' endpoint to retrieve a list of recorder logs based on the specified query parameters.
}
\examples{
# Retrieve a dataframe of recorder logs for project '017_neeri'
\dontrun{
get_recorderlogs(project_name = "pam_in_chemnitz")
}

}
