% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recorders}
\alias{get_recorders}
\title{Get recorder log info.}
\usage{
get_recorders(...)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/v0.1/docs/#operation/recorders_list}.}

\item{recorder_name}{The name of the recorder to retrieve information about.}
}
\value{
A list containing information about the specified recorder log: \url{https://api.ecopi.de/api/docs/#operation/recorderlogs_read}.

A dataframe containing the recorders that match the specified query parameters: \url{https://api.ecopi.de/api/v0.1/docs/#operation/recorders_list}.
}
\description{
Wrapper around the 'recorderlogs_read' endpoint to retrieve information about a specific recorder log.
}
\details{
This function retrieves information about a specific recorder log, based on the recorder name provided in the 'recorder_name' parameter.
The available information about a recorder log is documented in the EcoPi API documentation: \url{https://api.ecopi.de/api/docs/#operation/recorderlogs_read}.
}
\examples{
# Retrieve information about the recorder log for recorder '00000000d76d0bf9'
get_recorder_log_info(recorder_name = "00000000d76d0bf9")

# Retrieve a list of recorders for project '017_neeri'
get_recorders(project_name = "017_neeri")

}
