% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recorders_count_detections}
\alias{get_recorders_count_detections}
\title{Get a count of detections per recorder grouped by given parameters.}
\usage{
get_recorders_count_detections(project_name, ...)
}
\arguments{
\item{project_name}{Name of the project to get a count on detetcions class and recorder}

\item{...}{query paramaters. See \url{http://192.168.10.30:8001/api/docs/#tag/v0.2/operation/v0.2_aggregations_project_recorder_recordings_count_detections_retrieve}}
}
\value{
A summary containing species counts per recorder within a project that match the specified query parameters: \url{http://192.168.10.30:8001/api/docs/#tag/v0.2/operation/v0.2_aggregations_project_recorder_recordings_count_detections_retrieve}.
}
\description{
Wrapper around the 'aggregations_project_recorders_count_detections' endpoint to retrieve a count of detections for each recorder and grouped by the specified query parameters.
}
\examples{
# Retrieve a count pre species and recorders. By default, the count is returned for today
\dontrun{
get_recorders_count_detections(project_name = "039_zitro")
}
# You can specify a start_date and end_date to get a count for specific time interval.
# Limit 'countable' detections by setting a confidence threshold
\dontrun{
get_recorders_count_detections(project_name = "039_zitro", start_datetime = "2023-01-01", end_datetime = "2023-12-31", min_confidence = 0.85)
}

}
