% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recorderspeciescounts}
\alias{get_recorderspeciescounts}
\title{Get summary files of detections.}
\usage{
get_recorderspeciescounts(project_name, ...)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/v0.1/docs/#tag/meta}}
}
\value{
A summary containing species counts per recorder within a project that match the specified query parameters: \url{https://api.ecopi.de/api/v0.1/docs/#tag/meta}.
}
\description{
Wrapper around the 'Summary of species counts per recorder within a project' endpoint to retrieve a count of species detections for each recorder in a given project.
}
\examples{
# Retrieve a summary of species counts per recorder within a project
get_recorderstates()

# Retrieve a list of recorder states for recorder '00000000d76d0bf9'
get_summary(project_name = "039_zitro")

}
