% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recorderstates}
\alias{get_recorderstates}
\title{Get recorder states list.}
\usage{
get_recorderstates(..., get_response = FALSE)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/docs/#tag/v0.2/operation/v0.2_recorderstates_list}}

\item{get_response}{Logical. If TRUE, returns a list containing both the original response and the data frame.}
}
\value{
If get_response is FALSE (default), returns a data frame containing the recorder states.
If get_response is TRUE, returns a list with 'data' (the data frame) and 'response' (the original API response).
}
\description{
Wrapper around the 'recorderstates_list' endpoint to retrieve a list of recorder states based on the specified query parameters.
}
\examples{
# Retrieve all recorder states
\dontrun{
get_recorderstates()
}

# Retrieve a list of recorder states for recorder '00000000d76d0bf9'
\dontrun{
get_recorderstates(recorder_name = "00000000d76d0bf9")
}

}
