% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{patch_detections}
\alias{patch_detections}
\title{PATCH detection}
\usage{
patch_detections(..., id_or_uid, file_path)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/docs/#tag/v0.2/operation/v0.2_detections_partial_update}}

\item{id_or_uid}{The database ID or UID of the respective detection}

\item{file_path}{Path to file to upload (Watchout! If file_path given, new_data gets ignored (cannot patch multiple types)).}
}
\value{
httr2_response
}
\description{
Wrapper around the 'detections_partial_update' endpoint to update detections parameters based on the specified query parameters.
}
\examples{
# Update the parameter confirmed of an example detection
\dontrun{
patch_detections(id_or_uid = "64733fbc-7cc8-49f6-adf1-c9ec2d676959", confirmed = "YES")
}

}
