% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{patch_detections}
\alias{patch_detections}
\title{PATCH detection}
\usage{
patch_detections(..., id_or_uid)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/docs/#tag/v0.1/operation/v0.1_detections_partial_update}}

\item{id_or_uid}{The database ID or UID of the respective detection}
}
\value{
httr2_response
}
\description{
Wrapper around the 'detections_partial_update' endpoint to update detections parameters based on the specified query parameters.
}
\examples{
# Update the parameter confirmed of an example detection
\dontrun{patch_detections(id_or_uid = "64733fbc-7cc8-49f6-adf1-c9ec2d676959", confirmed = "YES")}

}
