% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{patch_recorders}
\alias{patch_recorders}
\title{Get recorder info.}
\usage{
patch_recorders(..., recorder_name, new_data)
}
\arguments{
\item{...}{query paramaters. See \url{https://api.ecopi.de/api/v0.1/docs/#operation/recorders_partial_update}.}

\item{recorder_name}{The name of the recorder to update information from.}
}
\value{
A list containing information about the specified recorder: \url{https://api.ecopi.de/api/docs/#operation/recorders_read}.

httr2_response
}
\description{
Wrapper around the 'recorders_read' endpoint to retrieve information about a specific recorder.
}
\details{
This function retrieves information about a specific recorder, based on the recorder name provided in the 'recorder_name' parameter.
The available information about a recorder is documented in the EcoPi API documentation: \url{https://api.ecopi.de/api/docs/#operation/recorders_read}.
}
\examples{
# Retrieve information about the recorder '00000000d76d0bf9'
get_recorderinfo(recorder_name = "00000000d76d0bf9")

# Update the parameter description of the recorder 00041aefd7jgg1014
patch_recorders(recorder_name = "00041aefd7jgg1014", description= "This a recorder ...", lat= 48)

}
